refactor: rename bind login request and update dialog state name, change popup window size
